+++ src/webgpu/index.js
@@ -1,2 +1,5 @@
 export * from './WebGPUPathTracer.js';
 export * from './BlurredEnvMapGenerator.js';
+
+import './shims/EquirectCameraShim.js';
+import './shims/PhysicalCameraShim.js';
